requirements.txt
pyobjc-framework-EventKit>=10.0
google-api-python-client>=2.0
google-auth-oauthlib>=1.0
google-auth-httplib2>=0.2
